Move login page back to the actual page
